Implementa melhorias na busca de livros, permite que o usuário escolha qual livro cadastrar. Utiliza id do gutendex para evitar duplicidade de livros

+++ src/main/java/br/com/alura/literalura/model/DadosLivro.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 
-public record DadosLivro(@JsonAlias("id") Integer id,
+public record DadosLivro(@JsonAlias("id") Long id,
                          @JsonAlias("title") String titulo,
                          @JsonAlias("authors") List<DadosAutor> autores,
                          @JsonAlias("languages") List<String> idiomas,
